+++ CHANGES.rst
@@ -4,9 +4,15 @@ Changelog
 1.5 (unreleased)
 ----------------
 
-New:
+Incompatibilities:
 
 - *add item here*
+
+New:
+
+- Add header 'Incompatibilities:' in postrelease.  In prerelease check
+  if at least one of New, Fixes, Incompatibilities is there.
+  See https://github.com/plone/Products.CMFPlone/issues/1323  [maurits]
 
 Fixes: