+++ CHANGES.rst
@@ -17,6 +17,7 @@ Bug fixes:
 - Fix new versions if they had dev/post release suffix.
   [gforcada]
 
+
 1.5.1 (2016-04-28)
 ------------------