+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.5.2 (unreleased)
+1.5.2 (2016-06-12)
 ------------------
 
 New features: